requirements.txt
# ZOONIVERSE CLIENT
panoptes_client==1.0

# DJANGO RELATED
django-countries==4.6.1
django==1.11.5
django-filter>=1.0.1
pytz==2017.2
six==1.11.0
mysqlclient==1.3.12
djangorestframework==3.7.3
markdown>=2.1.0

# Data import
django-import-export==1.0.0

# CORE ASTRO
astropy==2.0.2
numpy==1.13.1
pyvo==0.6.1
astroquery==0.3.5

# PLOTTING
matplotlib==2.2.2
#vaex==1.0.0b8
bokeh==0.12.15
django-toolbelt
Pillow==4.3.0
matplotlib==2.2.2

#For dev only:
django-extensions==1.9.7
ipython==5.6.0

future
pandas
scipy>=0.17.0

# Object recommender
git+https://github.com/chengsoonong/acton.git
protobuf
h5py
sklearn

#GPy
